ML-and-DA/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.28.2
pydantic==1.10.7
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
huggingface_hub==0.14.1
pandas==1.5.3
openpyxl==3.1.2
